minor changes after review
