+++ src/App.tsx
@@ -1,6 +1,6 @@
 import { Switch, Route, BrowserRouter as Router } from 'react-router-dom';
 
-import Header from './Header';
+import Header from './Header/Header';
 import Home from './Home';
 import About from './About';
 import Projects from './Projects';
